+++ src/bootstrap.ml
@@ -1,0 +1,1 @@
+let data_only_path _ = false