+++ src/stats.mli
@@ -5,3 +5,7 @@ val enable : unit -> unit
 
 (** If stats recording is enabled, collect stats now *)
 val record : unit -> unit
+
+val catapult : Catapult.t
+
+val enable_catapult : string -> unit